Fix sign conversion warnings in ODBC data binders
Add explicit static_cast<size_t>() for SQLLEN to size_t conversions when
calling resize() and constructing string views. This fixes -Wsign-conversion
warnings from clang when using pedantic compiler settings.

Files changed:
- BasicStringBinder.hpp: Multiple signedness casts in string operations
- SqlBinary.hpp: Size casts in binary data handling
- SqlDateTime.hpp: Cast duration count to unsigned for modulo operation
- StdString.hpp: Cast indicator to size_t in Resize()
- SqlError.hpp: Cast message length to size_t

+++ src/Lightweight/DataBinder/SqlDateTime.hpp
@@ -192,7 +192,7 @@ struct SqlDateTime
             .hour = (SQLUSMALLINT) hms.hours().count(),
             .minute = (SQLUSMALLINT) hms.minutes().count(),
             .second = (SQLUSMALLINT) hms.seconds().count(),
-            .fraction = (SQLUINTEGER) (((std::chrono::duration_cast<std::chrono::nanoseconds>(hms.to_duration()).count() % 1'000'000'000LLU) / 100) * 100)
+            .fraction = (SQLUINTEGER) (((static_cast<unsigned long long>(std::chrono::duration_cast<std::chrono::nanoseconds>(hms.to_duration()).count()) % 1'000'000'000LLU) / 100) * 100)
         };
         // clang-format on
     }
